release 3.1.37 source code for net

+++ Services/RocketMQ/V2/Model/ConsumerList.cs
@@ -26,7 +26,7 @@ namespace HuaweiCloud.SDK.RocketMQ.V2.Model
         /// Topic总数（当查询topic消费“列表”时才显示此参数）。
         /// </summary>
         [JsonProperty("total", NullValueHandling = NullValueHandling.Ignore)]
-        public decimal? Total { get; set; }
+        public int? Total { get; set; }
